Added piece images, working on list representation of board
Having difficulties representing the game board as a 2D list. I currently have a 2D list with the specified amount of rows and cols in the game board. Problem is i cant seem to populate the list wiht 1s for reds, 0s for black, and -1s for empty cells. I aim to implement that. Also i need to figure out how to represent the game pieces.