Working on piece movement
I cannot seem to make this player movement work, running into a lot of problems with the game logic and i don't know what or where. Begining to be a little frustrated. Need to take a step back, branch off from this and revert the feature branch to the most recent stable version